+++ src/models/DemoModel.js
@@ -1,0 +1,7 @@
+class DemoModel {
+    constructor(id, name){
+        this.id = id;
+        this.name = name;
+    }
+}
+export default User;
