+++ src/pages/home/HomeMainContent.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import PostCardList from '../PostCardList';
+import PostCardList from '../../components/PostCardList';
 import { Link } from "react-router-dom";
-import TopButton from '../TopButton';
+import TopButton from '../../components/TopButton';
 
 function MainContent() {
