+++ src/pages/post/post.js
@@ -1,8 +1,8 @@
 import React from "react";
 
 import PostContent from "./postContent";
-import Footer from "../Footer";
-import Navbar from "../Navbar";
+import Footer from "../../components/Footer";
+import Navbar from "../../components/Navbar";
 
 const Post = () =>{
     return(
