+++ src/pages/post/postContent.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import TopButton from '../TopButton';
+import TopButton from '../../components/TopButton';
 
 
 const PostContent = () => {
